+++ src/pygeoprocessing/geoprocessing.py
@@ -2104,10 +2104,10 @@ def merge_rasters(
         gtiff_creation_options=_DEFAULT_GTIFF_CREATION_OPTIONS):
     """Merge the given rasters into a single raster.
 
-    This operation calculates the bounding box of all the rasters in
-    `raster_path_list`, generates a new file of that size with the same pixel
-    band, and projection and copies the contents of each raster into the
-    correct georeferenced target.
+    This operation creates a mosaic of the rasters in `raster_path_list`.
+    The result is a raster of the size of the union of the bounding box of
+    the inputs where the contents of each raster are copied into the correct
+    georeferenced target.
 
     Note the input rasters must be in the same projection, same pixel size,
     same number of bands, and same datatype. If any of these are not true,